Moves Models to extra file ; codestyle
