+++ app/webFrontend/src/app/about/general-info/general-info.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {Contributor} from './contributor.model';
 
 @Component({
   selector: 'app-general-info',
@@ -9,7 +10,8 @@ export class GeneralInfoComponent implements OnInit {
 
   allContributors: Contributor[];
 
-  constructor() { }
+  constructor() {
+  }
 
   ngOnInit() {
     this.getAllContributors();
@@ -30,42 +32,4 @@ export class GeneralInfoComponent implements OnInit {
     ].sort(Contributor.compare);
   }
 
-}
-
-export class Contributor {
-  firstName: string;
-  name: string;
-  from: string;
-  to: string;
-  githubId: string;
-
-  constructor(firstName: string, name: string, from: string, githubId: string = '', to: string = 'present') {
-    this.firstName = firstName;
-    this.name = name;
-    this.from = from;
-    this.githubId = githubId;
-    this.to = to;
-  }
-
-  // To be able to sort eg an array
-  public static compare(a: Contributor, b: Contributor): number {
-    let compare;
-
-    // Check for FROM
-    compare = a.from.localeCompare(b.from);
-    if (compare !== 0) {
-      return compare;
-    }
-
-    // Check for NAME
-    compare = a.name.localeCompare(b.name);
-    if (compare !== 0) {
-      return compare;
-    }
-
-    // Check for FIRST_NAME
-    compare = a.firstName.localeCompare(b.firstName);
-
-    return compare;
-  }
 }
