+++ app/webFrontend/src/app/about/licenses/licenses.component.html
@@ -4,7 +4,9 @@
 <p *ngIf="allFrontendDependencies.length === 0">No Dependencies loaded...</p>
 <md-list>
   <md-list-item *ngFor="let dep of allFrontendDependencies;">
-    <h3 md-line>{{dep.name}} <small>{{dep.version}}</small></h3>
+    <h3 md-line>{{dep.name}}
+      <small>{{dep.version}}</small>
+    </h3>
     <p md-line>License: {{dep.summary}}</p>
   </md-list-item>
 </md-list>
@@ -13,7 +15,9 @@
 <p *ngIf="allApiDependencies.length === 0">No Dependencies loaded...</p>
 <md-list>
   <md-list-item *ngFor="let dep of allApiDependencies;">
-    <h3 md-line>{{dep.name}} <small>{{dep.version}}</small></h3>
+    <h3 md-line>{{dep.name}}
+      <small>{{dep.version}}</small>
+    </h3>
     <p md-line>License: {{dep.summary}}</p>
   </md-list-item>
 </md-list>
